New javascript text editor

+++ message.php
@@ -8,17 +8,39 @@ include 'message.inc.php';
 
 use \maierlabs\lpfw\Appl as Appl;
 
-Appl::addCss('editor/ui/trumbowyg.min.css');
-Appl::addJs('editor/trumbowyg.min.js');
-Appl::addJs('editor/langs/hu.min.js');
-Appl::addJsScript("
+\maierlabs\lpfw\Appl::addCss('editor/ui/trumbowyg.min.css');
+\maierlabs\lpfw\Appl::addJs('editor/trumbowyg.min.js');
+\maierlabs\lpfw\Appl::addCss('editor/plugins/specialchars/ui/trumbowyg.specialchars.min.css');
+\maierlabs\lpfw\Appl::addJs('editor/plugins/specialchars/trumbowyg.specialchars.min.js');
+\maierlabs\lpfw\Appl::addCss('editor/plugins/table/ui/trumbowyg.table.min.css');
+\maierlabs\lpfw\Appl::addJs('editor/plugins/table/trumbowyg.table.min.js');
+\maierlabs\lpfw\Appl::addCss('editor/plugins/colors/ui/trumbowyg.colors.min.css');
+\maierlabs\lpfw\Appl::addJs('editor/plugins/colors/trumbowyg.colors.min.js');
+\maierlabs\lpfw\Appl::addCss('editor/plugins/emoji/ui/trumbowyg.emoji.min.css');
+\maierlabs\lpfw\Appl::addJs('editor/plugins/emoji/trumbowyg.emoji.js');
+\maierlabs\lpfw\Appl::addJs('editor/plugins/pasteimage/trumbowyg.pasteimage.min.js');
+\maierlabs\lpfw\Appl::addJs('editor/langs/hu.min.js');
+\maierlabs\lpfw\Appl::addJsScript("
 	$( document ).ready(function() {
 		$('#story').trumbowyg({
 			fullscreenable: false,
 			closable: false,
 			lang: 'hu',
-			btns: ['formatting','btnGrp-design','|', 'link', 'insertImage','btnGrp-lists'],
+			btns: [
+                ['undo', 'redo'],
+                ['formatting'],
+                ['strong', 'em'],
+                ['superscript', 'subscript'],
+                ['link','insertImage','table','emoji'],
+                ['justifyLeft', 'justifyCenter', 'justifyRight'],
+                ['unorderedList', 'orderedList'],
+                ['horizontalRule'],
+                ['removeformat'],
+                ['specialChars'],
+			    ['foreColor', 'backColor'],['viewHTML']
+			],
 			removeformatPasted: true,
+			imageWidthModalEdit: true,
 			autogrow: true
 		});
 	});
@@ -31,18 +53,19 @@ if (isActionParam("postMessage")) {
 	if (userIsLoggedOn()) {
 		if (checkMessageContent($paramText)->ok) {
 			if (writeMessage($db,$paramText, getParam("privacy"), getLoggedInUserName($userDB))>=0) {
-				Appl::$resultDbOperation='<div class="alert alert-success" > A beadott üzenet elküldése sikerült!</div>';
+                Appl::setMessage("A beadott üzenet elküldése sikerült","success");
 				$paramName="";
 				$paramText="";
 			} else {
-				Appl::$resultDbOperation='<div class="alert alert-warning" > A beadott üzenet kimentése nem sikerült!</div>';
+                Appl::setMessage("A beadott üzenet kimentése nem sikerült!","warning");
 			}
-		} else 
-			Appl::$resultDbOperation='<div class="alert alert-warning" > A beadott üzenet úgytűnik nem tartalmaz érthető magyar szöveget! <br/> Probálkozz rövidítések nélkül vagy írj egy kicsitt bővebben.</div>';
+		} else {
+            Appl::setMessage("A beadott üzenet úgytűnik nem tartalmaz érthető magyar szöveget! <br/> Probálkozz rövidítések nélkül vagy írj egy kicsitt bővebben.", "info");
+        }
 	}
 	else {
 		if (strlen($paramName)<4) {
-			Appl::$resultDbOperation='<div class="alert alert-warning" >Írd be család és keresztneved!</div>';
+            Appl::setMessage("Írd be család és keresztneved!","warning");
 		}
 		else { 
 			if (checkMessageContent($paramText)->ok) {
@@ -75,14 +98,24 @@ if (isActionParam("checkMessage")) {
 }
 
 
-if (isActionParam("deleteMessage")) {
+if (isActionParam("deactivateMessage")) {
 	$id=getIntParam("id",-1);
 	if ($id!=-1) {
-		if (deleteMessage($db,$id))
+		if (deactivateMessage($db,$id))
             Appl::setMessage("Az üzenet ki lett törölve!","success");
 		else
             Appl::setMessage("Az üzenet törlése nem sikerült!","warning");
 	}
+}
+
+if (isActionParam("deleteMessage") && userIsAdmin()) {
+    $id=getIntParam("id",-1);
+    if ($id!=-1) {
+        if (deleteMessage($db,$id))
+            Appl::setMessage("Az üzenet véglegesen ki lett törölve!","success");
+        else
+            Appl::setMessage("Az üzenet végleges törlése nem sikerült!","warning");
+    }
 }
 
 if (isActionParam("commentMessage") && userIsAdmin()) {
@@ -95,14 +128,13 @@ if (isActionParam("commentMessage") && userIsAdmin()) {
 
 if (isActionParam("setPersonID") && userIsAdmin()) {
 	if ($db->saveMessagePersonID(getIntParam("id"),getParam("personid"))===true) {
-        Appl::setMessage("Személy cserérés sikerült.","success");
+        Appl::setMessage("Személy csere sikerült.","success");
 	} else {
-        Appl::setMessage("Személy csrélés nem sikerült!","warning");
+        Appl::setMessage("Személy csere nem sikerült!","warning");
 	}
 }
 
-Appl::setSiteTitle(getAktSchoolName()." véndiákok üzenőfala");
-Appl::setSiteSubTitle('Üzenőfal');
+Appl::setSiteTitle(getAktSchoolName()." véndiákok üzenőfala",'Üzenőfal');
 include("homemenu.inc.php");
 ?>
 
@@ -115,24 +147,24 @@ include("homemenu.inc.php");
   </button>
 </div>
 	<div id="message" style="margin-bottom: 10px;">
-		<form method="get" >
+		<form method="post" >
 			<?php if (!userIsLoggedOn()) {?>
-			<div class="input-group">
-				<span style="min-width:120px; text-align:right" class="input-group-addon" id="basic-addon1">Név</span>
-				<input type="text" class="form-control input-lg" value="<?php echo($paramName)?>" name="name" placeholder="családnév keresztnév"/>
-			</div>
+                <div class="input-group">
+                    <span style="min-width:120px; text-align:right" class="input-group-addon" id="basic-addon1">Név</span>
+                    <input type="text" class="form-control input-lg" value="<?php echo($paramName)?>" name="name" placeholder="családnév keresztnév"/>
+                </div>
 			<?php } ?>
-			<?php $text="~" ?>
-<div style="margin:1px"><textarea id="story" name="T" onchange="textChanged();" ><?php echo($paramText);?></textarea></div>
+			<?php $text="~~" ?>
+<div style="margin:1px"><textarea id="story" name="T" onchange="textChanged();"><?php echo(htmlspecialchars_decode($paramText));?></textarea></div>
 			<?php if (userIsLoggedOn()) {?>
-			<div class="radiogroup">
-				<div style="display: inline-block; padding:5px" >Ki láthatja<br /> ezt az üzenetet?</div>
-				<div title="Az egész világ" class="cradio radio_world"><input type="radio" name="privacy" value="world" <?php echo getFieldCheckedWord($text)?> onclick="saveMessage();" /></div>
-				<div title="Az iskolatársak" class="cradio radio_scool"><input type="radio" name="privacy" value="scool" <?php echo getFieldCheckedScool($text)?> onclick="saveMessage();" /></div>
-				<div title="Az osztálytársak" class="cradio radio_class"><input type="radio" name="privacy" value="class" <?php echo getFieldCheckedClass($text)?> onclick="saveMessage();" /></div>
-			</div> 
+                <div class="radiogroup">
+                    <div style="display: inline-block; padding:5px" >Ki láthatja<br /> ezt az üzenetet?</div>
+                    <div title="Az egész világ" class="cradio radio_world"><input type="radio" name="privacy" value="world" <?php echo getFieldCheckedWord($text)?> onclick="saveMessage();" /></div>
+                    <div title="Az iskolatársak" class="cradio radio_scool"><input type="radio" name="privacy" value="scool" <?php echo getFieldCheckedScool($text)?> onclick="saveMessage();" /></div>
+                    <div title="Az osztálytársak" class="cradio radio_class"><input type="radio" name="privacy" value="class" <?php echo getFieldCheckedClass($text)?> onclick="saveMessage();" /></div>
+                </div>
 			<?php } ?>
-			<button value="postMessage" name="action" class="btn btn-success" type="submit" ><span class="glyphicon glyphicon-send"></span> küldés!</button>
+			<button value="postMessage" name="action" class="btn btn-success" type="submit" ><span class="glyphicon glyphicon-send"></span> küldés</button>
 			<?php if (userIsAdmin()) {?>
 				<button value="checkMessage" name="action" class="btn btn-info" type="submit" ><span class="glyphicon glyphicon-check"></span> magyar?</button>
 			<?php } ?>
@@ -143,9 +175,7 @@ include("homemenu.inc.php");
 		</form>
 	</div>		
 
-<?php
-	displayMessageList(10);
-?>
+<?php displayMessageList(20); ?>
 	
 </div>
 <script type="text/javascript">
@@ -160,8 +190,12 @@ include("homemenu.inc.php");
 
     function deleteMessage(id) {
         if (confirm("Szeretnéd az üzenetet véglegesen törölni?")) {
-            $("#deleteId").val(id);
-            $("#deleteForm").submit();
+            document.location.href="message?id="+id+"&action=deleteMessage";
+        }
+    }
+    function deactivateMessage(id) {
+        if (confirm("Szeretnéd az üzenetet törölni?")) {
+            document.location.href="message?id="+id+"&action=deactivateMessage";
         }
     }
 </script>
